refactor: move preKeyBundleToUserClients to proteus dir
