+++ packages/core/src/messagingProtocols/proteus/ProteusService/Utility/getSessionsAndClientsFromRecipients/getSessionsAndClientsFromRecipients.test.ts
@@ -21,7 +21,7 @@ import type {UserClients} from '@wireapp/api-client/lib/conversation';
 import type {UserPreKeyBundleMap} from '@wireapp/api-client/lib/user';
 
 import {buildProteusService} from '../../../../../test/ProteusHelper';
-import {preKeyBundleToUserClients} from '../../../../../util/preKeyBundleToUserClients';
+import {preKeyBundleToUserClients} from '../preKeyBundleToUserClients';
 
 import {getSessionsAndClientsFromRecipients} from './';
